[Patch v6.9.35] Modularize features
